requirements.txt
flask>=3.0
opencv-python-headless>=4.9
mediapipe>=0.10.14
rembg[gpu]>=2.0.57
onnxruntime-gpu>=1.17
Pillow>=10.0
numpy>=1.26
